+++ app.js
@@ -812,6 +812,7 @@ import { renderers as RV1444 } from './views/group-v1444.js';
 import { renderers as RV1445 } from './views/group-v1445.js';
 import { renderers as RV1446 } from './views/group-v1446.js';
 import { renderers as RV1447 } from './views/group-v1447.js';
+import { renderers as RV1448 } from './views/group-v1448.js';
 import { renderers as RV1396 } from './views/group-v1396.js';
 import { renderers as RV164 } from './views/group-v164.js';
 import { renderers as RV165 } from './views/group-v165.js';
@@ -900,7 +901,7 @@ const RENDERERS = { ...RA, ...RB, ...RC, ...RE, ...RF, ...RG, ...RH, ...RI, ...R
   ...RV902,
   ...RV903,
   ...RV905,
-  ...RV906, ...RV907, ...RV908, ...RV909, ...RV910, ...RV911, ...RV912, ...RV916, ...RV917, ...RV918, ...RV919, ...RV920, ...RV921, ...RV922, ...RV923, ...RV924, ...RV925, ...RV927, ...RV928, ...RV932, ...RV958, ...RV960, ...RV1061, ...RV1062, ...RV1240, ...RV1241, ...RV1242, ...RV1243, ...RV1400, ...RV1389, ...RV1393, ...RV1397, ...RV1395, ...RV1390, ...RV1391, ...RV1392, ...RV1394, ...RV1399, ...RV1398, ...RV1401, ...RV1396, ...RV1412, ...RV1413, ...RV1414, ...RV1416, ...RV1417, ...RV1418, ...RV1419, ...RV1420, ...RV1421, ...RV1422, ...RV1423, ...RV1424, ...RV1425, ...RV1426, ...RV1427, ...RV1428, ...RV1429, ...RV1430, ...RV1431, ...RV1433, ...RV1434, ...RV1435, ...RV1436, ...RV1438, ...RV1439, ...RV1440, ...RV1441, ...RV1442, ...RV1443, ...RV1444, ...RV1445, ...RV1446, ...RV1447, ...RV63, ...RPALINT };
+  ...RV906, ...RV907, ...RV908, ...RV909, ...RV910, ...RV911, ...RV912, ...RV916, ...RV917, ...RV918, ...RV919, ...RV920, ...RV921, ...RV922, ...RV923, ...RV924, ...RV925, ...RV927, ...RV928, ...RV932, ...RV958, ...RV960, ...RV1061, ...RV1062, ...RV1240, ...RV1241, ...RV1242, ...RV1243, ...RV1400, ...RV1389, ...RV1393, ...RV1397, ...RV1395, ...RV1390, ...RV1391, ...RV1392, ...RV1394, ...RV1399, ...RV1398, ...RV1401, ...RV1396, ...RV1412, ...RV1413, ...RV1414, ...RV1416, ...RV1417, ...RV1418, ...RV1419, ...RV1420, ...RV1421, ...RV1422, ...RV1423, ...RV1424, ...RV1425, ...RV1426, ...RV1427, ...RV1428, ...RV1429, ...RV1430, ...RV1431, ...RV1433, ...RV1434, ...RV1435, ...RV1436, ...RV1438, ...RV1439, ...RV1440, ...RV1441, ...RV1442, ...RV1443, ...RV1444, ...RV1445, ...RV1446, ...RV1447, ...RV1448, ...RV63, ...RPALINT };
 
 // ----- Utility registry ----------------------------------------------------
 // Source of truth for routes, names, group, audiences, and clinical flag.
@@ -1562,6 +1563,7 @@ const UTILITIES = [
   { id: 'icdr-retinopathy',       name: 'Diabetic Retinopathy Severity (ICDR Scale)',        group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'hpa-glaucoma',           name: 'Hodapp-Parrish-Anderson Staging (Glaucoma Visual Field)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'isis-shoulder',          name: 'Instability Severity Index Score (Shoulder, ISIS)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
+  { id: 'abc-psi', name: 'ABC Classification (Posterior Shoulder Instability)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'snyder-slap', name: 'Snyder SLAP Lesion Classification', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'anaphylaxis-grade',      name: 'Anaphylaxis Severity Grade (Ring & Messmer)',       group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'anaphylaxis-criteria',   name: 'Anaphylaxis Diagnostic Criteria (2020 WAO)',        group: 'G', audiences: ['clinicians', 'educators'], clinical: true },